Big Trouble in little Decimal

+++ capstone/src/main/java/com/techelevator/view/Candy.java
@@ -1,7 +1,9 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Candy extends Product{
-    public Candy(String slotLocation, String name, double price, String category) {
+    public Candy(String slotLocation, String name, BigDecimal price, String category) {
         super(slotLocation, name, price, category);
     }
