+++ capstone/src/main/java/com/techelevator/view/Drink.java
@@ -1,7 +1,9 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Drink extends Product{
-    public Drink(String slotLocation, String name, double price, String category) {
+    public Drink(String slotLocation, String name, BigDecimal price, String category) {
         super(slotLocation, name, price, category);
     }
