+++ capstone/src/main/java/com/techelevator/view/Gum.java
@@ -1,7 +1,9 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Gum extends Product{
-    public Gum(String slotLocation, String name, double price, String category) {
+    public Gum(String slotLocation, String name, BigDecimal price, String category) {
         super(slotLocation, name, price, category);
     }
